creado SelectStatMultiplier.kt. Refactorizados algunos nombres de composables de HeroDuelScreen.kt.
